% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/average_partial_effect.R
\name{average_partial_effect}
\alias{average_partial_effect}
\title{Estimate average partial effects using a causal forest}
\usage{
average_partial_effect(forest)
}
\arguments{
\item{forest}{The trained forest.}
}
\value{
An estimate of the average partial effect, along with standard error.
}
\description{
Gets estimates of the average partial effect,
}
\details{
The (conditional) average treatment effect (target.sample = all):
  1/n sum_{i = 1}^n Cov[Wi, Yi | X = Xi] / Var[Wi | X = Xi].

Note that for a binary unconfounded treatment, the
average partial effect matches the average treatment effect.
}
\examples{
\dontrun{
n = 2000; p = 10
X = matrix(rnorm(n*p), n, p)
W = rbinom(n, 1, 1/(1 + exp(-X[,2]))) + rnorm(n)
Y = pmax(X[,1], 0) * W + X[,2] + pmin(X[,3], 0) + rnorm(n)
tau.forest = causal_forest(X, Y, W)
tau.hat = predict(tau.forest, X.test)
average_partial_effect(tau.forest)
}

}
