% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_average_treatment.R
\name{estimate_rate}
\alias{estimate_rate}
\title{Compute rate estimates, a function to be passed on to bootstrap routine.}
\usage{
estimate_rate(data, indices, q, wtd.mean)
}
\arguments{
\item{data}{A data.frame with the original data, column 1: DR.scores*sample.weights, column 2: sample.weights (positive),
column 3: priority scores (integer vector with scores 1,...,num.unique.prios).}

\item{indices}{A vector of indices which define the bootstrap sample.}

\item{q}{A vector of fractions to compute TOC on, with last entry = 1.}

\item{wtd.mean}{A weighting function determining the type of RATE estimate.}
}
\value{
an estimate of RATE, together with the TOC curve.
}
\description{
Compute rate estimates, a function to be passed on to bootstrap routine.
}
\keyword{internal}
