% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/causal_survival_forest.R
\name{expected_survival}
\alias{expected_survival}
\title{Compute E[T | X]}
\usage{
expected_survival(S.hat, Y.grid)
}
\arguments{
\item{S.hat}{The estimated survival curve.}

\item{Y.grid}{The time values corresponding to S.hat.}
}
\value{
A vector of expected values.
}
\description{
Compute E[T | X]
}
\keyword{internal}
