% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{export_graphviz}
\alias{export_graphviz}
\title{Export a tree in DOT format.
This function generates a GraphViz representation of the tree,
which is then written into `dot_string`.}
\usage{
export_graphviz(tree, include.na.path)
}
\arguments{
\item{tree}{the tree to convert}

\item{include.na.path}{A boolean toggling whether to include the path of missing values or not.}
}
\description{
Export a tree in DOT format.
This function generates a GraphViz representation of the tree,
which is then written into `dot_string`.
}
\keyword{internal}
