% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores}
\alias{get_scores}
\title{Compute doubly robust scores for a GRF forest object}
\usage{
get_scores(forest, ...)
}
\arguments{
\item{forest}{A grf forest object}

\item{...}{Additional arguments}
}
\value{
A vector of scores
}
\description{
Compute doubly robust scores for a GRF forest object
}
