% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores_CSF}
\alias{get_scores_CSF}
\title{Compute doubly robust for a causal survival forest.}
\usage{
get_scores_CSF(forest, subset = NULL)
}
\arguments{
\item{forest}{A trained causal survival forest.}

\item{subset}{Specifies subset of the training examples over which we
estimate the ATE. WARNING: For valid statistical performance,
the subset should be defined only using features Xi, not using
the treatment Wi or the outcome Yi.}
}
\description{
Compute doubly robust for a causal survival forest.
}
