% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{get_tree}
\alias{get_tree}
\title{Retrieve a single tree from a trained forest object.}
\usage{
get_tree(forest, index)
}
\arguments{
\item{forest}{The trained forest.}

\item{index}{The index of the tree to retrieve.}
}
\value{
A GRF tree object.
}
\description{
Retrieve a single tree from a trained forest object.
}
