% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{leaf_stats.causal_forest}
\alias{leaf_stats.causal_forest}
\title{Calculate summary stats given a set of samples for causal forests.}
\usage{
\method{leaf_stats}{causal_forest}(forest, samples, ...)
}
\arguments{
\item{forest}{The GRF forest}

\item{samples}{The samples to include in the calculations.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A named vector containing summary stats
}
\description{
Calculate summary stats given a set of samples for causal forests.
}
