% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{leaf_stats.default}
\alias{leaf_stats.default}
\title{A default leaf_stats for forests classes without a leaf_stats method
that always returns NULL.}
\usage{
\method{leaf_stats}{default}(forest, samples, ...)
}
\arguments{
\item{forest}{Any forest}

\item{samples}{The samples to include in the calculations.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
A default leaf_stats for forests classes without a leaf_stats method
that always returns NULL.
}
