% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{load_grf}
\alias{load_grf}
\title{Load a grf forest}
\usage{
load_grf(file, ...)
}
\arguments{
\item{file}{A connection or the name of the file where the R object is saved to or read from.}

\item{...}{Additional arguments passed to readRDS}
}
\value{
A grf forest object
}
\description{
Load a grf forest
}
