% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot.R
\name{plot.grf_tree}
\alias{plot.grf_tree}
\title{Plot a GRF tree object.}
\usage{
\method{plot}{grf_tree}(x, ...)
}
\arguments{
\item{x}{The tree to plot}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Plot a GRF tree object.
}
