% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_average_treatment.R
\name{plot.rank_average_treatment_effect}
\alias{plot.rank_average_treatment_effect}
\title{Plot the Targeting Operator Characteristic curve.}
\usage{
\method{plot}{rank_average_treatment_effect}(x, confidence.bars = TRUE, ...)
}
\arguments{
\item{x}{The output of rank_average_treatment_effect.}

\item{confidence.bars}{Whether to plot 95 \% confidence lines or not. Default is TRUE.}

\item{...}{Additional overridable arguments passed to plot.default.}
}
\description{
Plot the Targeting Operator Characteristic curve.
}
