% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/regression.forest.R
\name{predict.regression.forest}
\alias{predict.regression.forest}
\title{Predict with a regression forest}
\usage{
\method{predict}{regression.forest}(forest, newdata = NULL,
  num.threads = NULL, estimate.variance = FALSE)
}
\arguments{
\item{forest}{The trained forest.}

\item{newdata}{Points at which predictions should be made. If NULL,
makes out-of-bag predictions on the training set instead
(i.e., provides predictions at Xi using only trees that did
not use the i-th training example).}

\item{num.threads}{Number of threads used in training. If set to NULL, the software
automatically selects an appropriate amount.}

\item{estimate.variance}{Whether variance estimates for hat{tau}(x) are desired
(for confidence intervals).}
}
\value{
Vector of predictions.
}
\description{
Gets estimates of E[Y|X=x] using a trained regression forest.
}
