% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.boosted_regression_forest}
\alias{print.boosted_regression_forest}
\title{Print a boosted regression forest}
\usage{
\method{print}{boosted_regression_forest}(x, ...)
}
\arguments{
\item{x}{The boosted forest to print.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a boosted regression forest
}
