% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.grf}
\alias{print.grf}
\title{Print a GRF forest object.}
\usage{
\method{print}{grf}(x, decay.exponent = 2, max.depth = 4, ...)
}
\arguments{
\item{x}{The tree to print.}

\item{decay.exponent}{Controls the relative importance of splits.}

\item{max.depth}{The maximum depth of splits to consider.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a GRF forest object.
}
