% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.grf_tree}
\alias{print.grf_tree}
\title{Print a GRF tree object.}
\usage{
\method{print}{grf_tree}(x, ...)
}
\arguments{
\item{x}{The tree to print.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print a GRF tree object.
}
