% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.rank_average_treatment_effect}
\alias{print.rank_average_treatment_effect}
\title{Print the Rank-Weighted Average Treatment Effect (RATE).}
\usage{
\method{print}{rank_average_treatment_effect}(x, ...)
}
\arguments{
\item{x}{The output of rank_average_treatment_effect.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print the Rank-Weighted Average Treatment Effect (RATE).
}
