% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/print.R
\name{print.tuning_output}
\alias{print.tuning_output}
\title{Print tuning output.
Displays average error for q-quantiles of tuned parameters.}
\usage{
\method{print}{tuning_output}(x, tuning.quantiles = seq(0, 1, 0.2), ...)
}
\arguments{
\item{x}{The tuning output to print.}

\item{tuning.quantiles}{vector of quantiles to display average error over.
Default: seq(0, 1, 0.2) (quintiles)}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Print tuning output.
Displays average error for q-quantiles of tuned parameters.
}
