% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rd_forest.R
\name{rd_forest}
\alias{rd_forest}
\title{Regression discontinuity forest}
\usage{
rd_forest(
  X,
  Y,
  W,
  running,
  cutoff = 0,
  Y.hat = NULL,
  W.hat = NULL,
  Z.hat = NULL,
  num.trees = 2000,
  sample.weights = NULL,
  clusters = NULL,
  equalize.cluster.weights = FALSE,
  sample.fraction = 0.5,
  mtry = min(ceiling(sqrt(ncol(X)) + 20), ncol(X)),
  min.node.size = 5,
  honesty = TRUE,
  honesty.fraction = 0.5,
  honesty.prune.leaves = TRUE,
  alpha = 0.05,
  imbalance.penalty = 0,
  stabilize.splits = TRUE,
  ci.group.size = 2,
  reduced.form.weight = 0,
  tune.parameters = "none",
  tune.num.trees = 200,
  tune.num.reps = 50,
  tune.num.draws = 1000,
  compute.oob.predictions = TRUE,
  num.threads = NULL,
  seed = runif(1, 0, .Machine$integer.max)
)
}
\arguments{
\item{X}{The covariates used in the instrumental regression.}

\item{Y}{The outcome.}

\item{W}{The treatment received (may be binary or real).}

\item{running}{The running variable used to define the cutoff assignment.}

\item{cutoff}{Treatment threshold applied to the running variable. Default is 0.}

\item{Y.hat}{Estimates of the expected responses E[Y | Xi], marginalizing
over treatment. If Y.hat = NULL, these are estimated using
a separate regression forest. Default is NULL.}

\item{W.hat}{Estimates of the treatment propensities E[W | Xi]. If W.hat = NULL,
these are estimated using a separate regression forest. Default is NULL.}

\item{Z.hat}{Estimates of the instrument propensities E[Z | Xi]. If Z.hat = NULL,
these are estimated using a separate regression forest. Default is NULL.}

\item{num.trees}{Number of trees grown in the forest. Default is 2000.}

\item{sample.weights}{Weights given to each observation in estimation.
If NULL, each observation receives equal weight. Default is NULL.}

\item{clusters}{Vector of integers or factors specifying which cluster each observation corresponds to.
Default is NULL (ignored).}

\item{equalize.cluster.weights}{If FALSE, each unit is given the same weight (so that bigger
clusters get more weight). If TRUE, each cluster is given equal weight in the forest.
Default is FALSE.}

\item{sample.fraction}{Fraction of the data used to build each tree. Default is 0.5.}

\item{mtry}{Number of variables tried for each split. Default is sqrt(p) + 20.}

\item{min.node.size}{A target for the minimum number of observations in each tree leaf. Default is 5.}

\item{honesty}{Whether to use honest splitting. Default is TRUE.}

\item{honesty.fraction}{Fraction of data used for determining splits if honesty = TRUE. Default is 0.5.}

\item{honesty.prune.leaves}{Whether to prune empty leaves when honesty is TRUE. Default is TRUE.}

\item{alpha}{Maximum imbalance of a split. Default is 0.05.}

\item{imbalance.penalty}{Penalty for imbalanced splits. Default is 0.}

\item{stabilize.splits}{Whether or not the instrument should be taken into account when
determining the imbalance of a split. Default is TRUE.}

\item{ci.group.size}{The forest will grow ci.group.size trees on each subsample. Default is 2.}

\item{reduced.form.weight}{Whether splits should be regularized towards a naive
splitting criterion that ignores the instrument (and
instead emulates a causal forest). Default is 0.}

\item{tune.parameters}{Whether/how to tune parameters. Default is "none".}

\item{tune.num.trees}{Number of trees used for tuning. Default is 500.}

\item{tune.num.reps}{Number of random tuning draws. Default is 50.}

\item{tune.num.draws}{Number of subsamples used for tuning. Default is 1000.}

\item{compute.oob.predictions}{Whether OOB predictions on training set should be precomputed. Default is TRUE.}

\item{num.threads}{Number of threads used in training. Default is NULL.}

\item{seed}{Random seed. Default is a random draw.}
}
\value{
A trained rd_forest object.
}
\description{
Trains a regression discontinuity forest implemented as an instrumental forest where the
instrument is the cutoff indicator \eqn{Z = 1\{running \ge cutoff\}}.
}
