% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{save_grf}
\alias{save_grf}
\title{Save a grf forest}
\usage{
save_grf(forest, file = "", ...)
}
\arguments{
\item{forest}{The grf forest}

\item{file}{A connection or the name of the file where the R object is saved to or read from.}

\item{...}{Additional arguments passed to saveRDS}
}
\description{
Save a grf forest
}
