% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{schoolrct}
\alias{schoolrct}
\title{Brazilian high school financial education RCT}
\format{
A data frame containing outcomes and pre-treatment
characteristics for each student.
}
\source{
American Economic Association replication archive
(Bruhn et al., 2016).
}
\usage{
schoolrct
}
\description{
This dataset provides student-level information from a randomized controlled trial
evaluating the effect of a high school financial education program in Brazil.
Schools were randomly assigned to the program, and post-treatment financial literacy
was measured.
The data includes information on student outcomes and a variety of pre-treatment
characteristics such as survey responses and indices measuring students' ability
to save and their financial autonomy.
}
\details{
Original data licensed under Creative Commons Attribution 4.0 (CC BY 4.0).
This version has been lightly processed for demonstration purposes.
}
\references{
Bruhn, Miriam, Luciana de Souza Leão, Arianna Legovini, Rogelio Marchetti, and Bilal Zia.
"The impact of high school financial education: Evidence from a large-scale evaluation in Brazil."
American Economic Journal: Applied Economics 8, no. 4 (2016).
}
\keyword{datasets}
