% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analysis_tools.R
\name{split_frequencies}
\alias{split_frequencies}
\title{Get summaries of which features the forest split on}
\usage{
split_frequencies(forest, max.depth = 4)
}
\arguments{
\item{forest}{The trained forest.}

\item{max.depth}{Maximum depth of splits to consider.}
}
\description{
Get summaries of which features the forest split on
}
