% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tune_instrumental_forest}
\alias{tune_instrumental_forest}
\title{Instrumental forest tuning (removed)}
\usage{
tune_instrumental_forest(X, Y, W, Z, Y.hat, W.hat, Z.hat, ...)
}
\arguments{
\item{X}{X}

\item{Y}{Y}

\item{W}{W}

\item{Z}{Z}

\item{Y.hat}{Y.hat}

\item{W.hat}{W.hat}

\item{Z.hat}{Z.hat}

\item{...}{Additional arguments (currently ignored).}
}
\value{
output
}
\description{
To tune a instrumental forest, see the function `instrumental_forest`
}
