% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/tune_ll_regression_forest2.R
\name{tune_ll_regression_forest2}
\alias{tune_ll_regression_forest2}
\title{Local linear forest tuning}
\usage{
tune_ll_regression_forest2(
  forest,
  linear.correction.variables = NULL,
  ll.elnet.alpha = NULL,
  ll.weight.penalty = FALSE,
  num.threads = NULL,
  lambda.path = NULL,
  thresh = 1e-07,
  maxit = 1e+05,
  ...
)
}
\arguments{
\item{forest}{The forest used for prediction.}

\item{linear.correction.variables}{Variables to use for local linear prediction. If left null,
all variables are used. Default is NULL.}

\item{ll.elnet.alpha}{The elastic net mixing parameter, a proportion bound within 0 and 1. The parameter \code{ll.elnet.alpha}
is defined as \eqn{\alpha} in the elastic net penalty parameterization
\deqn{(1-\alpha)/2||\beta||_2^2+\alpha||\beta||_1.} \code{alpha=1} is the lasso penalty,
and \code{alpha=0} the ridge penalty.}

\item{ll.weight.penalty}{Option to standardize ridge penalty by covariance (TRUE),
or penalize all covariates equally (FALSE). Defaults to FALSE.}

\item{num.threads}{Number of threads used in training. If set to NULL, the software
automatically selects an appropriate amount.}

\item{lambda.path}{Optional list of lambdas to use for cross-validation.}

\item{thresh}{Convergence threshold for coordinate descent. Each coordinate descent loop continues until the
maximum change in the objective after any coefficient update is less than \code{thresh} times
the null deviance.}

\item{maxit}{Maximum number of passes over the data for all lambda values.}
}
\value{
A list of lambdas tried, corresponding errors, and optimal ridge penalty lambda.
}
\description{
Finds the optimal ridge penalty for local linear prediction.
}
\keyword{internal}
