% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{average_late}
\alias{average_late}
\title{Average LATE (removed)}
\usage{
average_late(forest, ...)
}
\arguments{
\item{forest}{The forest}

\item{...}{Additional arguments (currently ignored).}
}
\value{
output
}
\description{
See the function `average_treatment_effect`
}
