% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{average_partial_effect}
\alias{average_partial_effect}
\title{Average partial effect (removed)}
\usage{
average_partial_effect(forest, ...)
}
\arguments{
\item{forest}{The forest}

\item{...}{Additional arguments (currently ignored).}
}
\value{
output
}
\description{
See the function `average_treatment_effect`
}
