% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_average_treatment.R
\name{boot}
\alias{boot}
\title{Simple clustered bootstrap
Adopted from Bryan Ripley's `boot` boostrap package}
\usage{
boot(data, statistic, R, clusters, half.sample = TRUE)
}
\arguments{
\item{x}{The output of rank_average_treatment_effect.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Simple clustered bootstrap
Adopted from Bryan Ripley's `boot` boostrap package
}
\references{
Angelo Canty and Brian Ripley (2021). boot: Bootstrap R (S-Plus) Functions. R package version 1.3-28.
}
\keyword{internal}
