% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{custom_forest}
\alias{custom_forest}
\title{Custom forest (removed)}
\usage{
custom_forest(X, Y, ...)
}
\arguments{
\item{X}{X}

\item{Y}{Y}

\item{...}{Additional arguments (currently ignored).}
}
\description{
To build a custom forest, see an existing simpler forest, like regression_forest,
for a development template.
}
