% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{get_sample_weights}
\alias{get_sample_weights}
\title{Retrieve forest weights (renamed to get_forest_weights)}
\usage{
get_sample_weights(forest, ...)
}
\arguments{
\item{forest}{The trained forest.}

\item{...}{Additional arguments (currently ignored).}
}
\description{
Retrieve forest weights (renamed to get_forest_weights)
}
