% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores.multi_arm_causal_forest}
\alias{get_scores.multi_arm_causal_forest}
\title{Compute doubly robust scores for a multi arm causal forest.}
\usage{
\method{get_scores}{multi_arm_causal_forest}(forest, subset = NULL, outcome = 1, ...)
}
\arguments{
\item{forest}{A trained multi arm causal forest.}

\item{subset}{Specifies subset of the training examples over which we
estimate the ATE. WARNING: For valid statistical performance,
the subset should be defined only using features Xi, not using
the treatment Wi or the outcome Yi.}

\item{outcome}{In the event the forest is trained with multiple outcomes Y,
a column number/name specifying the outcome of interest. Default is 1.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
A matrix of scores for each contrast.
}
\description{
Compute doubly robust (AIPW) scores for average treatment effect estimation
using a multi arm causal forest. Under regularity conditions, the average of the DR.scores
is an efficient estimate of the average treatment effect.
}
