% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores_APE}
\alias{get_scores_APE}
\title{Compute doubly robust scores for average partial effect estimation with continuous
treatment, using a causal forest.}
\usage{
get_scores_APE(
  forest,
  subset = NULL,
  debiasing.weights = NULL,
  num.trees.for.weights = 500
)
}
\arguments{
\item{forest}{A trained causal forest.}

\item{subset}{Specifies subset of the training examples over which we
estimate the ATE. WARNING: For valid statistical performance,
the subset should be defined only using features Xi, not using
the treatment Wi or the outcome Yi.}

\item{debiasing.weights}{A vector of length n (or the subset length) of debiasing weights.
If NULL (default) these are obtained by estimating Var[W | X = x] using a new forest.}

\item{num.trees.for.weights}{Number of trees used to estimate Var[W | X = x]. Note: this
              argument is only used when debiasing.weights = NULL.

 @references Graham, Bryan S., and Cristine Campos de Xavier Pinto. "Semiparametrically
            efficient estimation of the average linear regression function." arXiv preprint
            arXiv:1810.12511, 2018.}
}
\description{
Compute doubly robust scores for average partial effect estimation with continuous
treatment, using a causal forest.
}
\references{
Hirshberg, David A., and Stefan Wager. "Augmented minimax linear estimation."
            arXiv preprint arXiv:1712.00038, 2017.
}
