% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/get_scores.R
\name{get_scores_ATE}
\alias{get_scores_ATE}
\title{Compute doubly robust (AIPW) scores for average treatment effect estimation
using a causal forest. Under regularity conditions, the average of the DR.scores
is an efficient estimate of the average treatment effect.}
\usage{
get_scores_ATE(forest, subset, debiasing.weights = NULL)
}
\arguments{
\item{forest}{A trained causal forest.}

\item{subset}{Specifies subset of the training examples over which we
estimate the ATE. WARNING: For valid statistical performance,
the subset should be defined only using features Xi, not using
the treatment Wi or the outcome Yi.}

\item{debiasing.weights}{A vector of length n (or the subset length) of debiasing weights.
If NULL (default) they are obtained via inverse-propensity weighting.}
}
\description{
Compute doubly robust (AIPW) scores for average treatment effect estimation
using a causal forest. Under regularity conditions, the average of the DR.scores
is an efficient estimate of the average treatment effect.
}
