% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rank_average_treatment.R
\name{plot.rank_average_treatment_effect}
\alias{plot.rank_average_treatment_effect}
\title{Plot the Targeting Operator Characteristic curve.}
\usage{
\method{plot}{rank_average_treatment_effect}(
  x,
  lb = x$TOC$estimate - 1.96 * x$TOC$std.err,
  ub = x$TOC$estimate + 1.96 * x$TOC$std.err,
  ylim = if (ci.lines) c(min(lb), max(ub)) else NULL,
  xlab = "q",
  ylab = "",
  main = "Targeting Operator Characteristic",
  sub = if (ci.lines) "(95 \% confidence bars in dashed lines)" else NULL,
  ci.lines = TRUE,
  ci.lty = 2,
  ci.col = "black",
  ...
)
}
\arguments{
\item{x}{The output of rank_average_treatment_effect.}

\item{lb}{Lower confidence lines. Default is 95\%.}

\item{ub}{Upper confidence lines. Default is 95\%.}

\item{ylim}{The `ylim` argument to plot.default.}

\item{xlab}{The `xlab` argument to plot.default.}

\item{main}{The title argument `main` to plot.default.}

\item{sub}{The `sub` title argument  to plot.default.}

\item{ci.lines}{Whether to plot 95 \% confidence lines or not. Default is TRUE.}

\item{ci.lty}{The line type `lty` for confidence lines.}

\item{ci.col}{The line color for confidence lines.}

\item{...}{Additional arguments passed to plot.default.}
}
\description{
Plot the Targeting Operator Characteristic curve.
}
\examples{
\donttest{
n <- 500
p <- 5
X <- matrix(rnorm(n * p), n, p)
W <- rbinom(n, 1, 0.5)
Y <- pmax(X[, 1], 0) * W + X[, 2] + pmin(X[, 3], 0) + rnorm(n)
cf <- causal_forest(X, Y, W)
rate1 <- rank_average_treatment_effect(cf, runif(n))
rate2 <- rank_average_treatment_effect(cf, pmax(X[, 1], 0))

# Customize some plot options.
plot(rate1, col = "blue", ci.col = "red", ci.lty = 1, main = "TOC")

# Plot two TOC curves.
par(mfrow = c(2, 1))
plot(rate1, sub = "", main = "Rule 1")
plot(rate2, sub = "", main = "Rule 2")

# To customize plots further use the curve data.frame in the `TOC` attribute directly.
}

}
