% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/custom_forest.R
\name{predict.custom_forest}
\alias{predict.custom_forest}
\title{Predict with a custom forest.}
\usage{
\method{predict}{custom_forest}(object, newdata = NULL, num.threads = NULL,
  ...)
}
\arguments{
\item{object}{The trained forest.}

\item{newdata}{Points at which predictions should be made. If NULL,
makes out-of-bag predictions on the training set instead
(i.e., provides predictions at Xi using only trees that did
not use the i-th training example).}

\item{num.threads}{Number of threads used in training. If set to NULL, the software
automatically selects an appropriate amount.}

\item{...}{Additional arguments (currently ignored).}
}
\value{
Vector of predictions.
}
\description{
Predict with a custom forest.
}
