% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tune_causal_forest}
\alias{tune_causal_forest}
\title{Causal forest tuning (removed)}
\usage{
tune_causal_forest(X, Y, W, Y.hat, W.hat, ...)
}
\arguments{
\item{X}{X}

\item{Y}{Y}

\item{W}{W}

\item{Y.hat}{Y.hat}

\item{W.hat}{W.hat}

\item{...}{Additional arguments (currently ignored).}
}
\value{
output
}
\description{
To tune a causal forest, see the function `causal_forest`
}
