% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/deprecated.R
\name{tune_regression_forest}
\alias{tune_regression_forest}
\title{Regression forest tuning (removed)}
\usage{
tune_regression_forest(X, Y, ...)
}
\arguments{
\item{X}{X}

\item{Y}{Y}

\item{...}{Additional arguments (currently ignored).}
}
\value{
output
}
\description{
To tune a regression forest, see the function `regression_forest`
}
